Raise when Encountering Unexpected Responses
Throughout the library we handle the case where the API responds
with something other than what we expected.  Previously, this was
handled by returning the JSON that the API gave us, instead of doing
anything with it like we would have if it was what we expected it
would be.  However, this behavior could be problematic - if code
was written expecting an object back, and instead it received a dict,
it would break in non-obvious ways.  To solve this, we could raise
anytime an unexpected response is received, making it immediately
obvious what went wrong.

On the other hand, if the response was being discarded anyway, this
change would cause code to break that otherwise would have worked.

This should only ever happen if the library is not written correctly,
or is not up to date with the API.  In normal circumstances, these
errors should never be encountered.