Adds id_relationships and documents linode.objects.base
id_relationships are like relationships, but expose the related object
through an attribute name that matches the defined name minus the _id
suffix (thereby allowing the underlying id to still be accessible as
expected).  This can replace boilerplate @property definitions in
subclasses of Base.

Also added documentation to the Base class and other classes in that
file.