Added _populate to Base.__init__
The code pattern of

```
o = Object(self._client, result['id'])
o._populate(result)
```

was getting old.  Made constructors take an optional json field
that is used to call _populate on creation.  This replaces the
above pattern with a single line, the constructor, with the
response json as a third parameter.

On initial inspection, this appears to work.  I'm going to review
it more thoroughly before considering it a success.