Clean Up Mappings
Refactor away linode.mappings in favor of static methods on classes.
Mappings used to be responsible for divining a type from a prefixed ID
and creating an object magically with no type needed, but when this
feature was removed upstream it was left without a real purpose.  This
change attempts to move its responsibilities into Base and PaginatedList
to clean up old code that is no longer relevant.  Also moved
PaginatedList out of util since nothing else ever lived there.