Adds LinodeClient.load
This closes #61

Adds a "load" method to LinodeClient to immediately load a type by ID.
Accepts the type to load (a class from linode.objects) and the ID of the
instance we are loading (and optionally a parent_id if applicable).
Immediately fetches the object and returns the object fully-populated,
or raises if the object is not found.