Preserve raw JSON in objects
This addresses the need presented in #141

There is not practical way to serialize objects once processed by this
library, as it applies many helpful (but not necessarily easy to
reverse) transformations to the data it receives from the API.  To
address this, I am preserving the raw JSON response the API returned for
objects.  This JSON is accessible to clients (although currently a
private property of the objects), and is even sufficient to populate a
new object if desired.  However, this approach has a few drawbacks:

 * Changes made to the object's mutable fields are not present in the raw JSON
 * The raw JSON is only updated if the object is repopulated, making it
   possible for the data to be stale if an object is kept in memory for
   a long time

Feedback on this approach is appreciated.  I can take a stab at
reversing the `_populate` function to generate a similar dict, but I
don't know if that's valuable (and it would be significantly more
effort).