paginated_list: Fix the 'IndexError' exception when max_pages=0
This could happen for example if the image filter was invalid, i.e.
'Image.is_public == None' or the API key simply doesn't have
permissions to list private images.
Along with the fix a new test case class is introduced and intends to
test the very basic functionality of the PaginatedList class, e.g.
instantiation.

Signed-off-by: Erik Skultety <eskultet@redhat.com>