bug: Handle empty PaginatedList responses
Closes #191

If a filter is given to a collection that resulted in no matches, the
resulting API response would include `"pages": 0`, `"page": 0`, and
`"results": 0`.  Those API responses are fine, but this library didn't
handle them, assuming there's always be at least one result.

This change fixes that assumption and allows this library to properly
return empty PaginatedLists.