Added support for authorized_users to disk_create
For whatever reason, this function accepts `**stackscript_args` instead
of arbitrary `**kwargs` to include in the API request, so
`authorized_users` needs to be added to the function's signature.  This
is less than ideal, and in future I may change the signature of this
function to make it behave as most other functions in this library do.