bug: Fix "Install on Linode" example project
This was last updated in 2019, and since then the
`OAuthScopes.Linodes.create` was removed in favor of the more modern
`OAuthScopes.Linodes.read_write`.  In addition to this, I fixed some
formatting and added comments in `app.py` to hopefully make it more
clear what the function of each piece is.

:warning: This example application still does not work, as CloudFlare is
currently blocking automated requests to the token expiration endpoint
on login.linode.com; I will be trying to get this fixed, and am leaving
the call to expire the token in place since it _should_ work as intended
here.