new: Add GHA release automation workflow (#237)
## 📝 Description

This change adds a `publish-pypi` GitHub Actions workflow for automating
releases to PyPI on GitHub release.

Additionally, this change alters the setup.py `version` field to
dynamically resolve the package version from either the
`LINODE_SDK_VERSION` environment variable or the embedded
`baked_version` file (required for source distributions). This change
allows us to dynamically version the package at build-time, which is a
requirement for the GHA release automation.

The `PYPI_API_TOKEN` secret has been added to this repository in
preparation for this change.