fix: `ip_addresses_share` to remove shared IPs (#316)
## 📝 Description

Fix an issue in `ip_addresses_share` to allow it unshare IPs. 

In the API doc, entering an empty IP array is allowed and it means
removing the shared IPs from the Linode. In the previous implementation,
we always looked for `ips[0]` when building the request param. It
brought the index out of range error when an empty IP array is given.

## ✔️ How to Test

`tox`